Response to code reviews. Put all request options into 1 file. renamed sample to download_view_image.py. Comments clean up
