--- tableauserverclient/server/request_options_base.py
@@ -1,9 +1,0 @@
-import abc
-
-
-class RequestOptionsBase(object):
-    __metaclass__ = abc.ABCMeta
-
-    @abc.abstractmethod
-    def apply_query_params(self, url):
-        return
